requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
starlette-context
python-jose
python-multipart
bcrypt
